refactor(mcp): split client into transport and response modules
- Move `StdioMcpToolCaller` and `McpClientHandler` to `transport.rs`
- Extract `normalize_tool_result` into `response.rs`
- Document backend module-layout conventions in `backends/mod.rs`
- Point `AGENTS.md` at the new conventions for future backends
